change payload to get all iterations of the geojson simulations
